+++ core/lib/core.dart
@@ -1,0 +1,7 @@
+library core;
+
+/// A Calculator.
+class Calculator {
+  /// Returns [value] plus 1.
+  int addOne(int value) => value + 1;
+}
